Fix acceptance tests.
Fix Rack::TrailingSlash so that it only chops off the trailing slash for
GET requests. Since redirecting the URL itself makes a GET request, this
can be a problem if the original request was, say, a DELETE request
(since it may lead the user to a delete confirmation page instead of
actually deleting the resource).