Further improvements to the integration test setup
Stories under a feature group are not part of the name of the group now and thus are not
inherited by scenarios. In other words, when a scenario has errors, and when it is
printed, the name of the scenario will not include the story. This way it's actually
readable. This is possible by using a separate formatter to print out the scenarios.

No need to use integration environment with envjs since it doesn't require a separate
instance of our app

When we run the integration tests and get failures, the next time we run the tests, only
run the failed tests